# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is API-docs of backend server of Harmonica team.",
        "title": "Harmonium backend API",
        "contact": {},
        "version": "1.0"
    },
    "host": "https://harmoniums.ru",
    "basePath": "api/v1",
    "paths": {
...
        "/pins": {
            "get": {
                "description": "Get pins by page",
                "tags": [
                    "Pins"
                ],
                "summary": "Pins list",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page num from 0",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.FeedPins"
                        }
                    },
                    "400": {
                        "description": "Possible code responses: 4.",
                        "schema": {
                            "$ref": "#/definitions/errs.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Possible code responses: 11.",
                        "schema": {
                            "$ref": "#/definitions/errs.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create pin by description",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pins"
                ],
                "summary": "Create pin",
                "parameters": [
                    {
                        "type": "string",
                        "default": "session-token=",
                        "description": "session-token",
                        "name": "Cookie",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Pin information in json",
                        "name": "pin",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Pin image",
                        "name": "image",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.PinPageResponse"
                        }
                    },
                    "400": {
                        "description": "Possible code responses: 3, 4, 15, 18, 19.",
                        "schema": {
                            "$ref": "#/definitions/errs.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Possible code responses: 2.",
                        "schema": {
                            "$ref": "#/definitions/errs.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Possible code responses: 11.",
                        "schema": {
                            "$ref": "#/definitions/errs.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/pins/created/{nickname}": {
            "get": {
                "description": "Get pins of user by page",
                "tags": [
                    "Pins"
                ],
                "summary": "Get pins that created by user id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page num from 0",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.FeedPins"
                        }
                    },
                    "400": {
                        "description": "Possible code responses: 4, 12.",
                        "schema": {
                            "$ref": "#/definitions/errs.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Possible code responses: 11.",
                        "schema": {
                            "$ref": "#/definitions/errs.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/pins/{pin_id}": {
            "get": {
                "description": "Get pin by id in the slug",
                "tags": [
                    "Pins"
                ],
                "summary": "Get pin by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Pin ID",
                        "name": "pin_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.PinPageResponse"
                        }
                    },
                    "400": {
                        "description": "Possible code responses: 4, 12.",
                        "schema": {
                            "$ref": "#/definitions/errs.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Possible code responses: 11.",
                        "schema": {
                            "$ref": "#/definitions/errs.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Update pin by description",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pins"
                ],
                "summary": "Update pin",
                "parameters": [
                    {
                        "description": "Pin information",
                        "name": "pin",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.Pin"
                        }
                    },
                    {
                        "type": "string",
                        "default": "session-token=",
                        "description": "session-token",
                        "name": "Cookie",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.PinPageResponse"
                        }
                    },
                    "400": {
                        "description": "Possible code responses: 3, 4, 12",
                        "schema": {
                            "$ref": "#/definitions/errs.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Possible code responses: 2.",
                        "schema": {
                            "$ref": "#/definitions/errs.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Possible code responses: 14.",
                        "schema": {
                            "$ref": "#/definitions/errs.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Possible code responses: 11.",
                        "schema": {
                            "$ref": "#/definitions/errs.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete pin by id (allowed only for pin's author)",
                "tags": [
                    "Pins"
                ],
                "summary": "Delete pin",
                "parameters": [
                    {
                        "type": "string",
                        "default": "session-token=",
                        "description": "session-token",
                        "name": "Cookie",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Pin ID",
                        "name": "pin_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.PinPageResponse"
                        }
                    },
                    "400": {
                        "description": "Possible code responses: 3, 4 12",
                        "schema": {
                            "$ref": "#/definitions/errs.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Possible code responses: 2.",
                        "schema": {
                            "$ref": "#/definitions/errs.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Possible code responses: 14.",
                        "schema": {
                            "$ref": "#/definitions/errs.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Possible code responses: 11.",
                        "schema": {
                            "$ref": "#/definitions/errs.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "entity.FeedPinResponse": {
            "description": "PinResponse information with author, pin id and content URL.",
            "type": "object",
            "properties": {
                "author": {
                    "$ref": "#/definitions/entity.PinAuthor"
                },
                "content_url": {
                    "type": "string"
                },
                "pin_id": {
                    "type": "integer"
                }
            }
        },
        "entity.FeedPins": {
            "description": "Pins array of FeedPinResponse",
            "type": "object",
            "properties": {
                "pins": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.FeedPinResponse"
                    }
                }
            }
        },
        "entity.Pin": {
            "description": "Full pin information",
            "type": "object",
            "properties": {
                "allow_comments": {
                    "type": "boolean"
                },
                "click_url": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "entity.PinAuthor": {
            "description": "User-author information with user id, nickname and avatar",
            "type": "object",
            "properties": {
                "avatar_url": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "entity.PinPageResponse": {
            "description": "Full pin information",
            "type": "object",
            "properties": {
                "allow_comments": {
                    "type": "boolean"
                },
                "author": {
                    "$ref": "#/definitions/entity.PinAuthor"
                },
                "click_url": {
                    "type": "string"
                },
                "content_url": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "is_owner": {
                    "type": "boolean"
                },
                "likes_count": {
                    "type": "integer"
                },
                "pin_id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "errs.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}